Switch inventory downloader from subclass
